Refactor tests to use a single file for each test case. Also helped fix some flakyness by adding permission handling. Noted that tests need to run without animations to work properly, so we should document this more heavily probably in README too. Made the GuiControls tests independent.
Otherwise the tests are the same as before, just more stable and laid out normally.
