+++ app/build.gradle
@@ -41,6 +41,9 @@ android {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+    testOptions {
+        animationsDisabled = true
+    }
     buildToolsVersion '28.0.3'
 }
