+++ app/src/androidTest/java/nz/org/cacophony/birdmonitor/DeleteRecordingsTest.java
@@ -1,52 +1,70 @@
 package nz.org.cacophony.birdmonitor;
 
 import android.content.Context;
+import android.support.test.espresso.IdlingRegistry;
+import android.support.test.filters.LargeTest;
 import android.support.test.rule.ActivityTestRule;
+import android.support.test.rule.GrantPermissionRule;
+import android.support.test.runner.AndroidJUnit4;
+import org.junit.*;
+import org.junit.runner.RunWith;
 
-import java.io.File;
-
+import static android.Manifest.permission.*;
 import static android.support.test.InstrumentationRegistry.getInstrumentation;
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.action.ViewActions.swipeLeft;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static nz.org.cacophony.birdmonitor.HelperCode.nowSwipeLeft;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-/**
- * Created by Tim Hunt on 16-Mar-18.
- */
+@LargeTest
+@RunWith(AndroidJUnit4.class)
+public class DeleteRecordingsTest {
 
-@SuppressWarnings("unused")
-class UploadRecordings {
+    private Context targetContext;
+    private Prefs prefs;
 
-    private static Context targetContext;
-    private static Prefs prefs;
-    private static File recordingsFolder;
-    private static File[] recordingFiles;
+    @Rule
+    public final ActivityTestRule<MainActivity> mActivityTestRule = new ActivityTestRule<>(MainActivity.class);
 
-    public static void uploadRecordings(ActivityTestRule<MainActivity> mActivityTestRule) {
+    @Rule
+    public GrantPermissionRule permissionRule = GrantPermissionRule.grant(
+            WRITE_EXTERNAL_STORAGE,
+            RECORD_AUDIO,
+            ACCESS_FINE_LOCATION,
+            READ_PHONE_STATE);
 
-        setUpForUploadAllRecordings(mActivityTestRule);
-        uploadAllRecordings();
-        tearDownForDeleteAllRecordings(mActivityTestRule);
+    @BeforeClass
+    public static void registerIdlingResource() {
+        IdlingRegistry.getInstance().register(IdlingResourceForEspressoTesting.recordIdlingResource);
+        IdlingRegistry.getInstance().register(IdlingResourceForEspressoTesting.signInIdlingResource);
     }
 
+    @AfterClass
+    public static void unregisterIdlingResource() {
+        IdlingRegistry.getInstance().unregister(IdlingResourceForEspressoTesting.recordIdlingResource);
+        IdlingRegistry.getInstance().unregister(IdlingResourceForEspressoTesting.signInIdlingResource);
+    }
 
-    private static void setUpForUploadAllRecordings(ActivityTestRule<MainActivity> mActivityTestRule) {
-
+    @Before
+    public void setUpForDeleteAllRecordings() {
         targetContext = getInstrumentation().getTargetContext();
         prefs = new Prefs(targetContext);
+
         prefs.setInternetConnectionMode("normal");
+
+        // Need to make sure app isn't disabled, so 'RECORD NOW' button is visible/enabled
+        prefs.setIsDisabled(true);
+
 
         if (prefs.getDeviceName() == null) {
             // Welcome Dialog WILL be displayed - and SetupWizard will be running
             HelperCode.dismissWelcomeDialog();
         } else {
             // Main menu will be showing
-
             onView(withId(R.id.btnSetup)).perform(click());
         }
 
@@ -58,24 +76,17 @@ class UploadRecordings {
 
         // Need to sign in
         HelperCode.signInUserTimhot();
-        try {
-            Thread.sleep(1000); // had to put in sleep, as could not work out how to consistently get groups to display before testing code tries to choose a group
-        } catch (Exception ignored) {
 
-        }
         nowSwipeLeft(); // takes you to Groups screen
 
         HelperCode.registerPhone(prefs);
 
         nowSwipeLeft(); // takes you to GPS
+        prefs.setIsDisabled(false);
         nowSwipeLeft(); // takes you to Test RecordAndSaveOnPhone
 
         // Need to put phone into offline mode so it doesn't try to upload the recording
         prefs.setInternetConnectionMode("offline");
-
-        prefs.setIsDisabled(false);
-
-        int numberOfRecordingsBeforeTestRecord = Util.getNumberOfRecordings(targetContext);
 
         onView(withId(R.id.btnRecordNow)).perform(click());
 
@@ -85,40 +96,27 @@ class UploadRecordings {
         onView(withId(R.id.btnFinished)).perform(click());
 
         onView(withId(R.id.btnAdvanced)).perform(click());
-
-        prefs.setLastRecordIdReturnedFromServer(-1);
-
     }
 
-
-    private static void tearDownForDeleteAllRecordings(ActivityTestRule<MainActivity> mActivityTestRule) {
+    @After
+    public void tearDownForDeleteAllRecordings() {
         prefs.setInternetConnectionMode("normal");
-        prefs.setIsDisabled(false);
         Util.signOutUser(targetContext);
+        prefs.setIsDisabled(false);
     }
 
 
-    private static void uploadAllRecordings() {
+    @Test
+    public void deleteRecordingsTest() {
+        int numberOfRecordingsBeforeDelete = Util.getNumberOfRecordings(targetContext);
+        assertTrue(numberOfRecordingsBeforeDelete > 0);
 
-        int numberOfRecordingsBeforeUpload = Util.getNumberOfRecordings(targetContext);
-        assertTrue(numberOfRecordingsBeforeUpload > 0);
-        prefs.setInternetConnectionMode("normal");
-
-        onView(withId(R.id.btnUploadFiles)).perform(click());
-
+        onView(withId(R.id.btnDeleteAllRecordings)).perform(click());
+        HelperCode.dismissDialogWithYes();
 
         int numberOfRecordingsAfterDelete = Util.getNumberOfRecordings(targetContext);
         assertEquals(0, numberOfRecordingsAfterDelete);
 
-        long lastRecordIdReturnedFromServer = prefs.getLastRecordIdReturnedFromServer();
-
-        assertTrue(lastRecordIdReturnedFromServer > 0);
-
-        onView(withId(R.id.tvMessagesManageRecordings)).check(matches(withText("Recordings have been uploaded to the server.")));
+        onView(withId(R.id.tvMessagesManageRecordings)).check(matches(withText("All recordings on the phone have been deleted.")));
     }
-
-    private static void nowSwipeLeft() {
-        onView(withId(R.id.SetUpWizard)).perform(swipeLeft());
-    }
-
 }
