--- app/src/androidTest/java/nz/org/cacophony/birdmonitor/RunGuiControls.java
@@ -1,47 +1,0 @@
-package nz.org.cacophony.birdmonitor;
-
-import android.support.test.filters.LargeTest;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
-
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-/**
- * Created by Tim Hunt on 14-Mar-18.
- */
-
-
-@LargeTest
-@RunWith(AndroidJUnit4.class)
-public class RunGuiControls {
-
-
-    @Rule
-    public final ActivityTestRule<MainActivity> mActivityTestRule = new ActivityTestRule<>(MainActivity.class);
-
-    @Before
-    public void registerIdlingResource() {
-        // let espresso know to synchronize with background tasks
-
-    }
-
-
-    @After
-    public void unregisterIdlingResource() {
-
-    }
-
-
-    @Test
-    public void guiControls() {
-
-        GuiControls.guiControls(mActivityTestRule);
-
-    }
-
-
-}
