+++ app/src/androidTest/java/nz/org/cacophony/birdmonitor/SignInUserTest.java
@@ -1,37 +1,47 @@
 package nz.org.cacophony.birdmonitor;
 
 import android.content.Context;
+import android.support.test.espresso.IdlingRegistry;
+import android.support.test.filters.LargeTest;
 import android.support.test.rule.ActivityTestRule;
+import android.support.test.runner.AndroidJUnit4;
+import org.junit.*;
+import org.junit.runner.RunWith;
 
 import static android.support.test.InstrumentationRegistry.getInstrumentation;
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.action.ViewActions.swipeLeft;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static nz.org.cacophony.birdmonitor.HelperCode.nowSwipeLeft;
 import static org.junit.Assert.assertTrue;
 
-/**
- * Created by Tim Hunt on 16-Mar-18.
- */
+@LargeTest
+@RunWith(AndroidJUnit4.class)
+public class SignInUserTest {
 
-@SuppressWarnings("unused")
-class SignInUser {
+    private Prefs prefs;
 
-    private static Prefs prefs;
+    @Rule
+    public final ActivityTestRule<MainActivity> mActivityTestRule = new ActivityTestRule<>(MainActivity.class);
 
-    public static void signInUser(ActivityTestRule<MainActivity> mActivityTestRule) {
-
-        setUpForSignInUser(mActivityTestRule);
-
-        signIn();
-
-        tearDownForSignInUser(mActivityTestRule);
+    @BeforeClass
+    public static void registerIdlingResource() {
+        IdlingRegistry.getInstance().register(IdlingResourceForEspressoTesting.createAccountIdlingResource);
+        IdlingRegistry.getInstance().register(IdlingResourceForEspressoTesting.registerPhoneIdlingResource);
+        IdlingRegistry.getInstance().register(IdlingResourceForEspressoTesting.signInIdlingResource);
     }
 
-    private static void setUpForSignInUser(ActivityTestRule<MainActivity> mActivityTestRule) {
+    @AfterClass
+    public static void unregisterIdlingResource() {
+        IdlingRegistry.getInstance().unregister(IdlingResourceForEspressoTesting.createAccountIdlingResource);
+        IdlingRegistry.getInstance().unregister(IdlingResourceForEspressoTesting.registerPhoneIdlingResource);
+        IdlingRegistry.getInstance().unregister(IdlingResourceForEspressoTesting.signInIdlingResource);
+    }
 
+    @Before
+    public void setUpForSignInUser() {
         Context targetContext = getInstrumentation().getTargetContext();
         prefs = new Prefs(targetContext);
         prefs.setInternetConnectionMode("normal");
@@ -50,28 +60,15 @@ class SignInUser {
 
         nowSwipeLeft();
         nowSwipeLeft(); // takes you to Sign In screen
-
     }
 
-    private static void tearDownForSignInUser(ActivityTestRule<MainActivity> mActivityTestRule) {
-
-    }
-
-
-    private static void signIn() {
-
+    @Test
+    public void signInTest() {
         HelperCode.signInUserTimhot();
 
         boolean userSignedIn = prefs.getUserSignedIn();
 
         assertTrue(userSignedIn);
         onView(withId(R.id.tvTitleMessageSignIn)).check(matches(withText("Signed In")));
-
-    }
-
-    private static void nowSwipeLeft() {
-
-        onView(withId(R.id.SetUpWizard)).perform(swipeLeft());
-
     }
 }
