+++ app/src/androidTest/java/nz/org/cacophony/birdmonitor/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * These tests are flaky at best with animations enabled due to the way Espresso works.
+ * Disable animations in the Android settings before running these tests.
+ */
+package nz.org.cacophony.birdmonitor;
